+++ Controller/Adminhtml/Landingpage/AbstractAction.php
@@ -5,13 +5,14 @@ namespace Algolia\AlgoliaSearch\Controller\Adminhtml\Landingpage;
 use Algolia\AlgoliaSearch\Helper\MerchandisingHelper;
 use Algolia\AlgoliaSearch\Model\LandingPageFactory;
 use Magento\Backend\App\Action\Context;
-use Magento\Framework\Registry;
+use Magento\Framework\Session\SessionManagerInterface;
 use Magento\Store\Model\StoreManagerInterface;
 
 abstract class AbstractAction extends \Magento\Backend\App\Action
 {
-    /** @var Registry */
-    protected $coreRegistry;
+
+    /** @var SessionManagerInterface */
+    protected $backendSession;
 
     /** @var LandingPageFactory */
     protected $landingPageFactory;
@@ -24,21 +25,21 @@ abstract class AbstractAction extends \Magento\Backend\App\Action
 
     /**
      * @param Context $context
-     * @param Registry $coreRegistry
+     * @param SessionManagerInterface $backendSession
      * @param LandingPageFactory $landingPageFactory
      * @param MerchandisingHelper $merchandisingHelper
      * @param StoreManagerInterface $storeManager
      */
     public function __construct(
         Context $context,
-        Registry $coreRegistry,
+        SessionManagerInterface $backendSession,
         LandingPageFactory $landingPageFactory,
         MerchandisingHelper $merchandisingHelper,
         StoreManagerInterface $storeManager
     ) {
         parent::__construct($context);
 
-        $this->coreRegistry = $coreRegistry;
+        $this->backendSession = $backendSession;
         $this->landingPageFactory = $landingPageFactory;
         $this->merchandisingHelper = $merchandisingHelper;
         $this->storeManager = $storeManager;
@@ -65,7 +66,7 @@ abstract class AbstractAction extends \Magento\Backend\App\Action
             }
         }
 
-        $this->coreRegistry->register('algoliasearch_landing_page', $landingPage);
+        $this->backendSession->setData('algoliasearch_landing_page', $landingPage);
 
         return $landingPage;
     }