+++ Helper/Entity/Product/PriceManager/Configurable.php
@@ -9,15 +9,20 @@ class Configurable extends ProductWithChildren
     /**
      * @param $groupId
      * @param $product
+     * @param $subProducts
      * @return float|int|mixed
      */
-    protected function getRulePrice($groupId, $product)
+    protected function getRulePrice($groupId, $product, $subProducts)
     {
         $childrenPrices = [];
-        /** @var \Magento\ConfigurableProduct\Model\Product\Type\Configurable $typeInstance */
         $typeInstance = $product->getTypeInstance();
-        $children = $typeInstance->getUsedProducts($product);
-        foreach ($children as $child) {
+
+        if (!$typeInstance instanceof \Magento\ConfigurableProduct\Model\Product\Type\Configurable) {
+            $this->logger->log('Unexpected product type encountered, reverting to default price calculation. Where Product Id is ' .$product->getId(). ' and Group Id is ' .$groupId);
+            return parent::getRulePrice($groupId, $product, $subProducts);
+        }
+
+        foreach ($subProducts as $child) {
             $childrenPrices[] = (float) $this->rule->getRulePrice(
                 new DateTime(),
                 $this->store->getWebsiteId(),